add route choice dataset, tests not working
